(chore) break up mono page into smaller components
