+++ src/types/agent.ts
@@ -1,18 +1,16 @@
 import * as z from "zod";
 
-
-
 export const generateImageSchema = z.object({
   project_title: z.string(),
   prompt: z.string(),
   // user_uname: z.string(),
-  user_id: z.number(),
+  user_id: z.string(),
 });
 
 export const regenerateImageSchema = z.object({
   project_title: z.string(),
   prompt: z.string(),
-  user_id: z.number(),
+  user_id: z.string(),
   guidance_scale: z.number(),
   negative_prompt: z.string(),
 });
@@ -25,7 +23,7 @@ export const createSurveySchema = z.object({
 
 export const createImageVariationSchema = z.object({
   guidance_prompt: z.string(),
-  user_id: z.number(),
+  user_id: z.string(),
   project_title: z.string(),
   input_image: z.string(),
 });
@@ -38,7 +36,7 @@ export const listPromptsSchema = z.object({
 
 export const pastGenerationSchema = z.object({
   id: z.number(),
-  user_id: z.number(),
+  user_id: z.string(),
   project_title: z.string(),
   prompt: z.string(),
   image_url: z.string(),
@@ -51,6 +49,3 @@ export type GenerateImageInput = z.infer<typeof generateImageSchema>;
 export type CreateImageInput = z.infer<typeof createImageVariationSchema>;
 export const pastGenerationsSchema = z.array(pastGenerationSchema);
 export type PastGenerations = z.infer<typeof pastGenerationsSchema>;
-
-
-
